requirements.txt
mne
tqdm
h5py
xlrd
numpy
scipy
pandas
scikit-learn
xgboost
seaborn
jupyter
openpyxl
kendall-w
quantities
matplotlib
joypy
ruamel.yaml
biosppy
hrv-analysis
streamlit
streamlit-aggrid
distinctipy
umap-learn
tensorboard
torchinfo
wandb
einops
torchmetrics
flirt
biopsykit
neurokit2
lightning
torch-tb-profiler
scikit-optimize